move workflow definition from dictionary to workflow
